Refactor: 파일 구조 개선 #22
