+++ src/components/features/date-detail/written-content.tsx
